Refactor revalidate route to receive validated GROQ-powered Webhook from Sanity.io.
