--- app/(site)/api/revalidate/route.ts
@@ -1,38 +1,0 @@
-import { revalidateSecret } from '@/sanity/sanity.api'
-import { revalidateTag } from 'next/cache'
-import { type NextRequest, NextResponse } from 'next/server'
-import { parseBody } from 'next-sanity/webhook'
-
-export async function POST(req: NextRequest) {
-  console.log(req.body);
-  try {
-    const { body, isValidSignature } = await parseBody<{
-      _type: string
-      slug?: string | undefined
-    }>(req, revalidateSecret)
-    if (!isValidSignature) {
-      const message = 'Invalid signature'
-      return new Response(message, { status: 401 })
-    }
-    console.log('body:', JSON.stringify(body, null, 2));
-
-    console.log(`isValidSignature: ${isValidSignature}`)
-    if (!body?._type) {
-      return new Response('Bad Request', { status: 400 })
-    }
-    console.log('Working!!!!!!!!')
-    revalidateTag(body._type)
-    if (body.slug) {
-      revalidateTag(`${body._type}:${body.slug}`)
-    }
-    return NextResponse.json({
-      status: 200,
-      revalidated: true,
-      now: Date.now(),
-      body,
-    })
-  } catch (err: any) {
-    console.error(err)
-    return new Response(err.message, { status: 500 })
-  }
-}
